Enable ODBC Build On Linux
- Enable Linux ODBC build with unicode support
- Added `docker-compose` for Flight SQL ODBC
- Register ODBC after build
- Replaced `boost::lexicographical_compare` with `std::lexicographical_compare`
- Fixed conversion bugs in `SetAttributeSQLWCHAR`
- Convert from std::string to std::u16string directly without involving wide string (wstring)
* Add patch for Windows build

+++ cpp/src/arrow/flight/sql/odbc/odbc_impl/accessors/string_array_accessor_test.cc
@@ -44,7 +44,7 @@ TEST(StringArrayAccessor, TestCDataTypeCharBasic) {
             accessor.GetColumnarData(&binding, 0, values.size(), value_offset, false,
                                      diagnostics, nullptr));
 
-  for (int i = 0; i < values.size(); ++i) {
+  for (size_t i = 0; i < values.size(); ++i) {
     ASSERT_EQ(values[i].length(), str_len_buffer[i]);
     ASSERT_EQ(values[i], std::string(buffer.data() + i * max_str_len));
   }
@@ -103,7 +103,7 @@ TEST(StringArrayAccessor, TestCDataTypeWcharBasic) {
             accessor->GetColumnarData(&binding, 0, values.size(), value_offset, false,
                                       diagnostics, nullptr));
 
-  for (int i = 0; i < values.size(); ++i) {
+  for (size_t i = 0; i < values.size(); ++i) {
     ASSERT_EQ(values[i].length() * GetSqlWCharSize(), str_len_buffer[i]);
     std::vector<uint8_t> expected;
     Utf8ToWcs(values[i].c_str(), &expected);
